@@ -10,7 +10,7 @@ import { provideRouter } from '@angular/router';
 import { RouterTestingHarness } from '@angular/router/testing';
 import { provideLocationMocks } from '@angular/common/testing';
 import { routes } from './app.routing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import type { WarehouseEngineStore } from './store/initial-state';
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
 import { provideMockActions } from '@ngrx/effects/testing';
@@ -19,6 +19,7 @@ import { ReplaySubject } from 'rxjs';
 import type { Action } from '@ngrx/store';
 import * as AuthActions from './store/auth/auth.actions';
 import { AuthEffects } from './store/auth/auth.effects';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 @Component({
   selector: 'app-top-nav, mat-spinner',
@@ -65,17 +66,19 @@ describe('AppComponent', () => {
     sessionStorage.clear();
 
     await TestBed.configureTestingModule({
-      providers: [
+    imports: [NoopAnimationsModule, AppComponent],
+    providers: [
         { provide: ThemeService, useValue: themeServiceSpy },
         provideRouter(routes),
         provideLocationMocks(),
         provideMockStore({ initialState }),
         provideMockActions(() => actions$),
         AuthEffects,
-        AuthFacade
-      ],
-      imports: [NoopAnimationsModule, AppComponent, HttpClientTestingModule],
-    })
+        AuthFacade,
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+})
       .overrideComponent(AppComponent, {
         add: {
           imports: [StubComponent, SideNavStubComponent, LoginMockComponent],
