@@ -1,19 +1,22 @@
 import { TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 
 import { LoginService } from './login.service';
 import { AuthFacade } from 'src/app/store/auth/auth.facade';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('LoginService', () => {
   let service: LoginService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      providers: [
-        AuthFacade
-      ],
-      imports: [HttpClientTestingModule],
-    });
+    imports: [],
+    providers: [
+        AuthFacade,
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+});
     service = TestBed.inject(LoginService);
   });
 
